+++ Magic/CHANGELOG.md
@@ -10,6 +10,7 @@
  - Add `brush_entities` parameter to engineering spells, to turn off replicate and clone copying entities
  - Add `clear_containers` parameter to engineering spells, to clear (the copied versions of) containers when copying them
  - Add `min_radius` parameter to AreaOfEffect action, to select entities within a ring
+ - Add `/mage discover` command for learning crafting recipes
 
 ## Changes